+++ packages/iov-utils/src/assert.ts
@@ -1,5 +1,0 @@
-export function assert(condition: any, msg?: string): asserts condition {
-  if (!condition) {
-    throw new Error(msg || "condition is not truthy");
-  }
-}